+++ junit-tests/src/test/java/org/junit/gen5/engine/junit5/resolver/ClassResolverTests.java
@@ -14,7 +14,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.gen5.engine.discovery.ClassSelector.forClass;
 import static org.junit.gen5.engine.junit5.resolver.ClassResolver.descriptorForParentAndClass;
 import static org.junit.gen5.engine.junit5.resolver.PackageResolver.descriptorForParentAndName;
-import static org.junit.gen5.launcher.main.DiscoveryRequestBuilder.request;
+import static org.junit.gen5.launcher.main.TestDiscoveryRequestBuilder.request;
 
 import java.util.Optional;
