+++ junit-tests/src/test/java/org/junit/gen5/engine/junit5/resolver/PackageResolverTests.java
@@ -13,7 +13,7 @@ package org.junit.gen5.engine.junit5.resolver;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.gen5.engine.discovery.PackageSelector.forPackageName;
 import static org.junit.gen5.engine.junit5.resolver.PackageResolver.descriptorForParentAndName;
-import static org.junit.gen5.launcher.main.DiscoveryRequestBuilder.request;
+import static org.junit.gen5.launcher.main.TestDiscoveryRequestBuilder.request;
 
 import java.util.Optional;
 import java.util.Set;
